+++ Skin/Retail/SharedXML/SharedColorConstants.lua
@@ -14,6 +14,9 @@ do --[[ SharedXML\SharedColorConstants.lua ]]
         Color.Create(0.7098, 0.5019, 0.1725),
     }
 
+    private.PAPER_FRAME_TITLE_COLOR = Color.white
+    private.PAPER_FRAME_TEXT_COLOR = Color.grayLight
+
     private.FACTION_COLORS = {
         Alliance = Color.Create(0.0, 0.2, 0.6),
         Horde = Color.Create(0.5, 0.0, 0.0),